project-board/src/main/java/com/personal/board/service/ArticleCommentService.java
package com.personal.board.service;

import com.personal.board.DTO.ArticleDto;
import com.personal.board.repository.ArticleCommentRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@RequiredArgsConstructor
@Transactional
@Service
public class ArticleCommentService {

    private final ArticleCommentRepository articleCommentRepository;

    @Transactional(readOnly = true)
    public List<ArticleDto> searchArticleComment(Long articleId) {
        return List.of();
    }
}
